pyproject.toml
[tool.poetry]
name = "polls"
version = "0.1.0"
description = ""
authors = ["remimd"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
blacksheep = "^1.2.8"
cryptography = "^38.0.4"
django = "^4.1.4"
django-extensions = "^3.2.1"
django-phonenumber-field = "^7.0.1"
phonenumbers = "^8.13.3"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.2"
pyjwt = "^2.6.0"
python-dotenv = "^0.21.0"
toml = "^0.10.2"
tzlocal = "^4.2"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-env = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
